Consolidate training scripts into trainer CLI
